daemon: move proxy-related code to `proxy.go`
Signed-off by: Ian Vernon <ian@cilium.io>
